asf/medical/api/auth.py
"""
Authentication module for the Medical Research Synthesizer API.

This module provides JWT-based authentication for the FastAPI implementation,
including user management, token generation, and validation.
"""

import os
from datetime import datetime, timedelta
from typing import Dict, Optional, List

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from jose import JWTError, jwt
from passlib.context import CryptContext
from pydantic import BaseModel

# Security configuration
SECRET_KEY = os.environ.get("JWT_SECRET_KEY", "super-secret-key-change-in-production")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60  # 1 hour

# Password hashing
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# OAuth2 scheme
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

# Models
class Token(BaseModel):
    access_token: str
    token_type: str
    role: str

class TokenData(BaseModel):
    email: Optional[str] = None

class User(BaseModel):
    email: str
    role: str

class UserInDB(User):
    hashed_password: str

# Mock user database (replace with a real database in production)
users_db = {
    "admin@example.com": {
        "hashed_password": pwd_context.hash("admin_password"),
        "role": "admin"
    },
    "user@example.com": {
        "hashed_password": pwd_context.hash("user_password"),
        "role": "user"
    }
}

# Helper functions
def verify_password(plain_password, hashed_password):
    """Verify a password against a hash."""
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password):
    """Generate a password hash."""
    return pwd_context.hash(password)

def get_user(db, email: str):
    """Get a user from the database."""
    if email in db:
        user_dict = db[email]
        return UserInDB(**user_dict, email=email)
    return None

def authenticate_user(db, email: str, password: str):
    """Authenticate a user."""
    user = get_user(db, email)
    if not user:
        return False
    if not verify_password(password, user.hashed_password):
        return False
    return user

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    """Create a JWT access token."""
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

async def get_current_user(token: str = Depends(oauth2_scheme)):
    """Get the current user from a JWT token."""
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        email: str = payload.get("sub")
        if email is None:
            raise credentials_exception
        token_data = TokenData(email=email)
    except JWTError:
        raise credentials_exception
    user = get_user(users_db, email=token_data.email)
    if user is None:
        raise credentials_exception
    return user

async def get_current_active_user(current_user: User = Depends(get_current_user)):
    """Get the current active user."""
    return current_user

def register_user(email: str, password: str, role: str = "user"):
    """Register a new user."""
    if email in users_db:
        return False
    users_db[email] = {
        "hashed_password": get_password_hash(password),
        "role": role
    }
    return True
