asf/medical/requirements.txt
# API
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.9
alembic>=1.10.3
psycopg2-binary>=2.9.6  # For PostgreSQL
asyncpg>=0.27.0  # For async PostgreSQL

# External APIs
httpx>=0.24.0
requests>=2.28.2
aiohttp>=3.8.4

# Data processing
pandas>=2.0.0
numpy>=1.24.2
scikit-learn>=1.2.2

# Export
fpdf>=1.7.2
openpyxl>=3.1.2
xlsxwriter>=3.0.9

# ML/AI
torch>=2.0.0
transformers>=4.28.1
shap>=0.41.0
geoopt>=0.5.0
faiss-cpu>=1.7.3  # Use faiss-gpu for GPU support

# Graph databases
neo4j>=5.7.0
pymemgraph>=0.1.0

# Distributed computing
ray>=2.3.1

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.2
loguru>=0.7.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0
